requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
aiohttp==3.8.6
apscheduler==3.10.4
python-hosts==1.0.3
python-multipart==0.0.6
jinja2==3.1.2
pyyaml==6.0.1
requests==2.31.0
transmission-rpc==2.0.0
croniter
dnspython==2.4.2
urllib3>=1.26.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
itsdangerous